[sip] Reorganize files to fit the optim plugin hierarchy
