+++ plugins/sip/main/optimizer.cpp
@@ -6,23 +6,23 @@
 //  Copyright © 2017 Antoine Marendet. All rights reserved.
 //
 
-#include "contractors/ibex_CtcBlankenship.h"
-#include "contractors/ibex_CtcEvaluation.h"
-#include "contractors/ibex_CtcFilterSICParameters.h"
-#include "contractors/ibex_CtcBisectActiveParameters.h"
-#include "contractors/ibex_CtcFirstOrderTest.h"
-#include "contractors/ibex_CtcHC4SIP.h"
-#include "contractors/ibex_GoldsztejnSICBisector.h"
-#include "loup/ibex_LoupFinderLineSearch.h"
-#include "loup/ibex_LoupFinderRestrictionsRelax.h"
-#include "loup/ibex_LoupFinderSIPDefault.h"
-#include "loup/ibex_LoupFinderCompo.h"
-#include "main/ibex_CellDoubleHeapSIP.h"
-#include "main/ibex_MinibexOptionsParser.h"
-#include "main/ibex_SIPOptimizer.h"
-#include "numeric/ibex_RelaxationLinearizerSIP.h"
-#include "numeric/ibex_RestrictionLinearizerSIP.h"
-#include "system/ibex_SIPSystem.h"
+#include "ibex_CtcBlankenship.h"
+#include "ibex_CtcEvaluation.h"
+#include "ibex_CtcFilterSICParameters.h"
+#include "ibex_CtcBisectActiveParameters.h"
+#include "ibex_CtcFirstOrderTest.h"
+#include "ibex_CtcHC4SIP.h"
+#include "ibex_GoldsztejnSICBisector.h"
+#include "ibex_LoupFinderLineSearch.h"
+#include "ibex_LoupFinderRestrictionsRelax.h"
+#include "ibex_LoupFinderSIPDefault.h"
+#include "ibex_LoupFinderCompo.h"
+#include "ibex_CellDoubleHeapSIP.h"
+#include "ibex_MinibexOptionsParser.h"
+#include "ibex_SIPOptimizer.h"
+#include "ibex_RelaxationLinearizerSIP.h"
+#include "ibex_RestrictionLinearizerSIP.h"
+#include "ibex_SIPSystem.h"
 
 #include "args.hxx"
 #include "ibex_CtcCompo.h"
